requirements.txt
python-dotenv>=0.19.0
# OpenAI (optional)
openai==0.28.0
# Google Cloud Vertex AI (Gemini) - pinned to compatible versions
google-cloud-aiplatform==1.63.0
vertexai==1.63.0
